wrap pto controller



# message auto-generated for no-merge-commit merge:
merge sjh_wrap_controller into master

wrap pto controller

Created-by: jhshen
Commit-by: jhshen95
Merged-by: DCGDDD
Description: 将pybind直接绑定的接口封装在pto_impl命名空间中，而对用户开放的接口封装在pto中

该PR完成对controller, enum, operation相关接口的封装修改

并在pybind中增加从int到SymbolicScalar的隐式类型转换

See merge request: cann/pypto-dev!115

+++ python/src/tensor_op/cpp_reference/src/dump_graph.cpp
@@ -26,7 +26,7 @@ using namespace npu::tile_fwk;
 int main() {
     Program::GetInstance().GetConfig().Reset();
 
-    std::vector<int> shape{1, 2, 256, 128, 2};
+    std::vector<int64_t> shape{1, 2, 256, 128, 2};
     Tensor a(DT_FP32, shape, "a");
     TileShape::Current().SetVecTile(1, 1, 32, 32, 2);
